Extracted utility methods into a separate static class
