+++ addons/Context/Tweek.Drivers.Context.MongoDb/MongoDbDriver.cs
@@ -1,20 +1,12 @@
 using FSharpUtils.Newtonsoft;
 using Newtonsoft.Json;
-using MongoDB;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Reflection;
 using System.Threading.Tasks;
-using LanguageExt;
 using Tweek.ApiService.Addons;
 using Tweek.Engine.DataTypes;
 using Tweek.Engine.Drivers.Context;
 using MongoDB.Driver;
-using MongoDB.Bson;
-using MongoDB.Bson.Serialization;
-using Newtonsoft.Json.Linq;
 
 namespace Tweek.Drivers.Context.MongoDb
 {
@@ -61,7 +53,7 @@ namespace Tweek.Drivers.Context.MongoDb
             var update = Builders<Dictionary<string, object>>.Update.Set("_id", id);
             foreach (var item in context)
             {
-                update = update.Set(item.Key, ConvertJsonValueToObject(item.Value));
+                update = update.Set(item.Key, ConversionUtils.ConvertJsonValueToObject(item.Value));
             }
             await collection.FindOneAndUpdateAsync(filter, update);
         }
@@ -72,7 +64,7 @@ namespace Tweek.Drivers.Context.MongoDb
             var collection = GetCollection(identity);
             var filter = Builders<Dictionary<string, object>>.Filter.Eq("_id", id);
             var result = await (await collection.FindAsync(filter)).SingleOrDefaultAsync();
-            return ToJsonValues(result);
+            return ConversionUtils.ToJsonValues(result);
         }
 
         public async Task RemoveFromContext(Identity identity, string key)
@@ -91,75 +83,11 @@ namespace Tweek.Drivers.Context.MongoDb
             {
                 var key = keyValuePair.Key;
                 var value = keyValuePair.Value;
-                result.Add(key, ConvertJsonValueToObject(value));
+                result.Add(key, ConversionUtils.ConvertJsonValueToObject(value));
             }
             
             return result;
         }
 
-        private static Dictionary<string, JsonValue> ToJsonValues(Dictionary<string, object> input)
-        {
-            var result = new Dictionary<string, JsonValue>(input.Count);
-            foreach (var keyValuePair in input)
-            {
-                var key = keyValuePair.Key;
-                if (key == "_id")
-                {
-                    continue;
-                }
-                var value = keyValuePair.Value;
-                result.Add(key, ConvertObjectToJsonValue(value));
-            }
-
-            return result;
-        }
-
-        private static object ConvertJsonValueToObject(JsonValue value)
-        {
-            if (value.IsNull)
-            {
-                return null;
-            }
-
-            if (value.IsBoolean)
-            {
-                return value.AsBoolean();
-            }
-
-            if (value.IsFloat || value.IsNumber)
-            {
-                return value.AsDecimal();
-            }
-            
-            if (value.IsString)
-            {
-                return value.AsString();
-            }
-            
-            if (value.IsArray)
-            {
-                return value.AsArray().Select(ConvertJsonValueToObject).ToArray();
-            }
-
-            if (!value.IsRecord)
-                throw new NotImplementedException($"There's not implementation for JsonValue tag {value.Tag}");
-            
-            var record = (JsonValue.Record) value;
-            return record.properties.ToDictionary(kvPair => kvPair.Item1, kvPair => ConvertJsonValueToObject(kvPair.Item2));
-        }
-
-        private static JsonValue ConvertObjectToJsonValue(object value)
-        {
-            return IsNumeric(value) ? JsonValue.NewNumber((decimal)value) : JsonValue.From(JToken.FromObject(value));
-        }
-
-        private static bool IsNumeric(object value)
-        {
-            var type = value.GetType();
-            return type == typeof(int) ||
-                   type == typeof(long) ||
-                   type == typeof(double) ||
-                   type == typeof(decimal);
-        }
     }
 }
