+++ services/authoring/package.json
@@ -37,7 +37,7 @@
     "multer": "^1.3.0",
     "nconf": "^0.9.1",
     "node-getopt": "^0.2.3",
-    "nodegit": "0.18.3",
+    "nodegit": "0.22.0",
     "passport": "^0.4.0",
     "passport-jwt": "^3.0.1",
     "passport-strategy": "^1.0.0",
